requirements.txt
openai
python-dotenv
sendgrid
feedparser
requests



